requirements.txt
google-api-python-client==2.14.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-cloud-storage==1.42.3
pandas==1.3.3
